django-team-tracker/requirements.txt
Django==2.2
django-filter==2.2.0
djangorestframework==3.10
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
